fix(cache): any go file inside a .gomake folder is added to the hash
Previously the cache busting hash only added the files that were parsed into an AST.
Now we consider any go file inside a .gomake folder, those closes #7

However keep in mind that changes in consumed packages outside of the .gomake folder will still not trigger a rebuild. In the future we may be able to make this smarter by traversing all imported packages. But this may not be performant, see how magefile does it.