fix(brew): must download a tarball and extract, can not download direct
Also changed the way scoop does it's install to use a zip file.

Also removed the direct binary downloads, it's probably unwise to download a binary directly and then just run it. At least if it's inside an archive of some sort then that forces to the user to have to perform additional work in order to use the program, which is a bit like saying, Are you really sure you want to run this?

Not only that the binary is compressed to half it's original size.